Refactor extensions to expose args
