+++ extensions/scarb-doc/src/lib.rs
@@ -1,8 +1,13 @@
+use crate::args::{Args, OutputFormat};
 use crate::db::ScarbDocDatabase;
+use crate::docs_generation::markdown::MarkdownContent;
+use crate::errors::MetadataCommandError;
 use crate::metadata::compilation::{
     crates_with_starknet, get_project_config, get_relevant_compilation_unit,
 };
-use anyhow::Result;
+use crate::metadata::get_target_dir;
+use crate::versioned_json_output::VersionedJsonOutput;
+use anyhow::{Context, Result, ensure};
 use cairo_lang_compiler::diagnostics::DiagnosticsReporter;
 use cairo_lang_diagnostics::{FormattedDiagnosticEntry, Severity};
 use cairo_lang_filesystem::{
@@ -11,12 +16,17 @@ use cairo_lang_filesystem::{
 };
 use errors::DiagnosticError;
 use itertools::Itertools;
-use scarb_metadata::{CompilationUnitMetadata, Metadata, PackageMetadata};
+use scarb_metadata::{
+    CompilationUnitMetadata, Metadata, MetadataCommand, PackageMetadata, ScarbCommand,
+};
 use scarb_ui::Ui;
+use scarb_ui::args::ToEnvVars;
+use scarb_ui::components::Status;
 use serde::Serialize;
 use smol_str::ToSmolStr;
 use types::Crate;
 
+pub mod args;
 pub mod db;
 pub mod docs_generation;
 pub mod errors;
@@ -35,6 +45,81 @@ pub struct PackageInformation {
 pub struct AdditionalMetadata {
     pub name: String,
     pub authors: Option<Vec<String>>,
+}
+
+const OUTPUT_DIR: &str = "doc";
+const JSON_OUTPUT_FILENAME: &str = "output.json";
+
+fn main_inner(args: Args, ui: Ui) -> Result<()> {
+    ensure!(
+        !args.build || matches!(args.output_format, OutputFormat::Markdown),
+        "`--build` is only supported for Markdown output format"
+    );
+    let metadata = MetadataCommand::new()
+        .inherit_stderr()
+        .envs(args.features.to_env_vars())
+        .exec()
+        .map_err(MetadataCommandError::from)?;
+    let metadata_for_packages = args.packages_filter.match_many(&metadata)?;
+    let output_dir = get_target_dir(&metadata).join(OUTPUT_DIR);
+
+    let packages_information = generate_packages_information(
+        &metadata,
+        &metadata_for_packages,
+        args.document_private_items,
+        ui.clone(),
+    )?;
+
+    match args.output_format {
+        OutputFormat::Json => {
+            VersionedJsonOutput::new(packages_information)
+                .save_to_file(&output_dir, JSON_OUTPUT_FILENAME)?;
+
+            let output_path = output_dir
+                .join(JSON_OUTPUT_FILENAME)
+                .strip_prefix(&metadata.workspace.root)
+                .unwrap_or(&output_dir)
+                .to_string();
+            ui.print(Status::new("Saving output to:", &output_path));
+        }
+        OutputFormat::Markdown => {
+            for pkg_information in packages_information {
+                let pkg_output_dir = output_dir.join(&pkg_information.metadata.name);
+
+                MarkdownContent::from_crate(&pkg_information)?
+                    .save(&pkg_output_dir)
+                    .with_context(|| {
+                        format!(
+                            "failed to save docs for package {}",
+                            pkg_information.metadata.name
+                        )
+                    })?;
+
+                let output_path = pkg_output_dir
+                    .strip_prefix(&metadata.workspace.root)
+                    .unwrap_or(&pkg_output_dir)
+                    .to_string();
+                ui.print(Status::new("Saving output to:", &output_path));
+                if args.build {
+                    let build_output_dir = pkg_output_dir.join("book");
+                    ScarbCommand::new()
+                        .arg("mdbook")
+                        .arg("--input")
+                        .arg(pkg_output_dir)
+                        .arg("--output")
+                        .arg(build_output_dir.clone())
+                        .env("SCARB_UI_VERBOSITY", ui.verbosity().to_string())
+                        .run()?;
+                    let output_path = build_output_dir
+                        .strip_prefix(&metadata.workspace.root)
+                        .unwrap_or(&build_output_dir)
+                        .to_string();
+                    ui.print(Status::new("Saving build output to:", &output_path));
+                }
+            }
+        }
+    }
+    Ok(())
 }
 
 pub fn generate_packages_information(
