+++ extensions/scarb-mdbook/src/main.rs
@@ -1,56 +1,22 @@
-use anyhow::{Context, Result};
-use camino::Utf8PathBuf;
 use clap::Parser;
-use mdbook::MDBook;
-use scarb_ui::Ui;
-use scarb_ui::args::VerbositySpec;
-use std::fs::File;
-use std::io::Write;
-use std::path::PathBuf;
+use scarb_mdbook::args::Args;
+use scarb_ui::{OutputFormat, Ui};
 use std::process::ExitCode;
-
-#[derive(Parser, Debug)]
-#[command(version, about, long_about = None)]
-struct Args {
-    /// Path to book source directory.
-    #[arg(long)]
-    pub input: Utf8PathBuf,
-    /// Path to book output directory.
-    #[arg(long)]
-    pub output: Utf8PathBuf,
-    /// Logging verbosity.
-    #[command(flatten)]
-    pub verbose: VerbositySpec,
-}
 
 fn main() -> ExitCode {
     let args = Args::parse();
-    let ui = Ui::new(args.verbose.clone().into(), scarb_ui::OutputFormat::Text);
-    let build_result = main_inner(&args, ui.clone()).with_context(|| {
+    let ui = Ui::new(args.verbose.clone().into(), OutputFormat::Text);
+    let build_result = scarb_mdbook::main_inner(&args, ui.clone()).map_err(|err| {
         format!(
-            "failed to build book from `{}` source path",
+            "failed to build book from `{}` source path: {err:#}",
             args.input.clone()
         )
     });
     match build_result {
         Ok(()) => ExitCode::SUCCESS,
         Err(error) => {
-            ui.error(format!("{error:#}"));
+            ui.error(error);
             ExitCode::FAILURE
         }
     }
-}
-
-fn main_inner(args: &Args, _ui: Ui) -> Result<()> {
-    let mut book = MDBook::load(args.input.clone())?;
-    let output_path: PathBuf = args.output.clone().into();
-    book.config.build.build_dir = output_path
-        .strip_prefix(&args.input)
-        .unwrap_or(&output_path.clone())
-        .into();
-    book.build()?;
-    let highlight = include_str!("../theme/highlight.js");
-    let mut highlight_file = File::create(output_path.join("highlight.js"))?;
-    highlight_file.write_all(highlight.as_bytes())?;
-    Ok(())
 }
